Co-localize blobs based on surrounding intensities
Implement an algorithm for co-localizing blobs based on surrounding intensities at each blob position in each channel. Thresholds for detection are first identified in each channel by taking the blobs in the given channel, finding the surrounding intensities, and taking a low (5th) percentile. Then for each channel, the surrounding intensities of blobs in that channel are compared with the thresholds in the other channels. Blobs exceeding any given threshold are considered to co-localize in that channel.