Option to calculate co-localization thresholds by minimum of surrounding blob intensities
Blob co-localization has taken a percentile of all pixels surrounding blobs in a given channel, which has required choosing this percentile. As another option, take the average surrounding intensity for each blob and set the minimum value as the threshold, which avoids the need to pick a percentile. This approach also ensures that blobs will at least co-localize with themselves.

Additionally:
- Add a parameter to customize the percentile when setting the threshold by a percentile
- Reuse the blobs ROI mask for each channel