Fix displaying blob co-localizations when showing blobs in select channels
Turning off channels removed blobs from those channels but did not removing the corresponding co-localizations, causing an exception when displaying these co-localizations alongside the blobs. Add a parameter to return the mask of blobs in the given channels and apply this mask to the co-localizations.

Also, turn off co-localizations when displaying blobs from a database since co-localizations are currently not stored in the database.