Change blob visibility and co-localization check boxes to lists
The `BooleanEditor` used for these boxes positioned the label on the left of the check box, which is inconsistent with the check boxes in the ROI panel since they use `CheckListEditor`s instead. For better consistency and the flexibility of adding more values, change these check boxes to list-based controls.