Create a storage class for image5d arrays
The main image is an `image5d` Numpy array, which has been stored in the `config` module along with a number of associated attributes. To begin encapsulating these attributes into a single object, create an `Image5d` storage class. The plan is to eventually migrate the `config` attributes related to `image5d` into this class, with the goal of allow multiple main images to be loaded in a single session.