Store loaded image5d paths in `Image5d` object
The path to save or load an imported image5d array archive has been constructed from a "base" path. The full path has not been saved, however, and typically the path is constructed during saving and reconstructed during loading. The GUI even deconstructs this path from a given full path. To make this path more accessible, store it in a `Image5d` instance when importing or loading an image5d archive, along with the associated metadata path. This object is not yet stored in `config` but allows the returned object from these functions to contain this data.