Use a common function for storing and loading experiment names in the database
Experiment name formatting has been performed independently in the GUI and for blob verifications. Use a common function instead, stored in the `sqlite` module since the name is currently used solely for database storage. Also, use the newly stored path of the loaded image rather than reconstructing the path from the base filename, which avoids assuming that the first sub-image has been loaded. Provide an option in the filename deconstruction function to preserve the sub-image since this string is used in experiment names for the database.