Provide an option to get the immediate parent of each label when exporting regions
Region export has used the `--labels level` parameter both to set the parent level and indicate whether to export all labels or only those in the atlas corresponding to the labels reference file. Change the latter flag to the `orig_colors` sub-argument, instead using a level of None to flag getting the immediate parent. Extend the parent retrieval function to get the immediate parent and also fix cases where no parents are found. Rename the `max_level` parameter in the function to find atlas labels since this function does not use the level directly, but only as an indicator to load the atlas labels.