Check for discrepancies between derived and listed parent in ontology references
The ABA ontology reference file lists the parent for each label, while the label parent parsing function derives these parents based on the given target level. Occasionally the found parents may be discrepant if the child is listed as the same level as that of the parent. Check for and report any of these differences, adding an enumeration for the ABA parent ID key.